requirements.txt
django-cors-headers==4.1.0
django-ninja==0.22.2
Django==4.2.2

